netlink: recreate udp connection when ipv4/ipv6 nlris are added
Creation of udp connection is invoked by encap nlri and it has
dependencies on ipv4/ipv6 nlris.
This means when encap nlri is advertised before dependent ipv4/ipv6
nlri, creation of udp connection fails.
To aid this, this patch tries to recreate udp connection when ipv4/ipv6
nlris are advertised.